Refactor scheduling-related middleware into separate module
