+++ src/controllers/group.js
@@ -1,35 +1,22 @@
-const moment = require('moment');
-const {
-  parseEventDates, eventProposal,
-} = require('../utils/event');
-
 // Model
 const User = require('../models/user');
-const PersonalSchedule = require('../models/personalSchedule');
 const UserGroup = require('../models/userGroup');
 const Group = require('../models/group');
-const GroupSchedule = require('../models/groupSchedule');
-const Post = require('../models/post');
-const PostDetail = require('../models/postDetail');
-const Comment = require('../models/comment');
 
 // Error
 const {
   DataFormatError, ApiError,
-  UserNotFoundError, UnathroizedError, ScheduleNotFoundError,
-  GroupNotFoundError, PostNotFoundError, EditPermissionError,
-  CommentNotFoundError, ExpiredCodeError, InvalidGroupJoinError,
+  UserNotFoundError, GroupNotFoundError,
+  UnauthorizedError, EditPermissionError,
+  ExpiredCodeError, InvalidGroupJoinError, UserIsLeaderError,
 } = require('../errors');
 
 // Validator
 const {
-  validateGroupSchema, validateGroupIdSchema, validateEventProposalSchema,
-  validateGroupScheduleSchema, validateScheduleDateScehma,
-  validatePostSchema, validatePostIdSchema, validatePageSchema,
-  validateCommentSchema, validateCommentIdSchema, validateGroupScheduleIdSchema,
+  validateGroupSchema, validateGroupIdSchema, validatePageSchema,
 } = require('../utils/validators');
 
-async function createGroup(req, res, next) {
+async function postGroup(req, res, next) {
   try {
     const { error: bodyError } = validateGroupSchema(req.body);
     if (bodyError) {
@@ -83,34 +70,26 @@ async function getGroupDetail(req, res, next) {
   }
 }
 
-async function deleteGroup(req, res, next) {
+async function getGroupList(req, res, next) {
   try {
-    const { error: paramError } = validateGroupIdSchema(req.params);
-    if (paramError) {
+    const { error: queryError } = validatePageSchema(req.query);
+    if (queryError) {
       return next(new DataFormatError());
     }
 
-    const { group_id: groupId } = req.params;
-    const [group, user] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    if (group.leader !== user.userId) {
-      return next(new UnathroizedError());
-    }
-
-    await group.destroy();
-    return res.status(204).json({ message: 'Successfully delete group' });
-  } catch (err) {
+    const { page } = req.query;
+    const pageSize = 9;
+    const startIndex = (page - 1) * pageSize;
+    const { rows } = await Group.findAndCountAll({
+      offset: startIndex,
+      limit: pageSize,
+    });
+    const result = rows.map((group) => ({
+      name: group.name,
+      member: group.member,
+    }));
+    return res.status(200).json(result);
+  } catch {
     return next(new ApiError());
   }
 }
@@ -150,6 +129,38 @@ async function patchGroup(req, res, next) {
   }
 }
 
+async function deleteGroup(req, res, next) {
+  try {
+    const { error: paramError } = validateGroupIdSchema(req.params);
+    if (paramError) {
+      return next(new DataFormatError());
+    }
+
+    const { group_id: groupId } = req.params;
+    const [group, user] = await Promise.all([
+      Group.findByPk(groupId),
+      User.findOne({ where: { nickname: req.nickname } }),
+    ]);
+
+    if (!group) {
+      return next(new GroupNotFoundError());
+    }
+
+    if (!user) {
+      return next(new UserNotFoundError());
+    }
+
+    if (group.leader !== user.userId) {
+      return next(new UnauthorizedError());
+    }
+
+    await group.destroy();
+    return res.status(204).json({ message: 'Successfully delete group' });
+  } catch (err) {
+    return next(new ApiError());
+  }
+}
+
 async function deleteGroupUser(req, res, next) {
   try {
     const { error: paramError } = validateGroupIdSchema(req.params);
@@ -171,7 +182,7 @@ async function deleteGroupUser(req, res, next) {
     }
 
     if (group.leader === user.userId) {
-      return next(new UnathroizedError());
+      return next(new UserIsLeaderError());
     }
 
     await UserGroup.destroy({
@@ -184,174 +195,6 @@ async function deleteGroupUser(req, res, next) {
     await group.update({ member: group.member - 1 });
 
     return res.status(204).json({ message: 'Successfully delete group user' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getGroupSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupIdSchema(req.params);
-    const { error: queryError } = validateScheduleDateScehma(req.query);
-    if (paramError || queryError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId } = req.params;
-    const group = await Group.findByPk(groupId);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    const { startDateTime, endDateTime } = req.query;
-    const start = moment.utc(startDateTime).toDate();
-    const end = moment.utc(endDateTime).toDate();
-    const groupEvent = await GroupSchedule.getSchedule([groupId], start, end);
-    const users = (await group.getUsers()).map((user) => user.userId);
-    const userEvent = await PersonalSchedule.getSchedule(users, start, end);
-    const event = {};
-    event.nonRecurrenceSchedule = [
-      ...userEvent.nonRecurrenceSchedule,
-      ...groupEvent.nonRecurrenceSchedule,
-    ];
-    event.recurrenceSchedule = [
-      ...userEvent.recurrenceSchedule,
-      ...groupEvent.recurrenceSchedule,
-    ];
-    return res.status(200).json(event);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function postGroupSchedule(req, res, next) {
-  try {
-    const { error: bodyError } = validateGroupScheduleSchema(req.body);
-    const { error: paramError } = validateGroupIdSchema(req.params);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId } = req.params;
-
-    const {
-      title,
-      content,
-      startDateTime,
-      endDateTime,
-      recurrence,
-      freq,
-      interval,
-      byweekday,
-      until,
-    } = req.body;
-
-    await GroupSchedule.create({
-      groupId,
-      title,
-      content,
-      startDateTime,
-      endDateTime,
-      recurrence,
-      freq,
-      interval,
-      byweekday,
-      until,
-      possible: null,
-      impossible: null,
-    });
-    return res.status(201).json({ message: 'Successfully create group schedule' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function putGroupSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupScheduleIdSchema(req.params);
-    const { error: bodyError } = validateGroupScheduleSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { schedule_id: scheduleId, group_id: groupId } = req.params;
-    const [user, group, schedule] = await Promise.all([
-      User.findOne({ where: { nickname: req.nickname } }),
-      Group.findByPk(groupId),
-      GroupSchedule.findByPk(scheduleId),
-    ]);
-
-    if (!schedule) {
-      return next(new ScheduleNotFoundError());
-    }
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (user.userId !== group.leader) {
-      return next(new EditPermissionError());
-    }
-
-    await schedule.update(req.body);
-    return res.status(201).json({ message: 'Successfully modify group schedule' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function deleteGroupSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupScheduleIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { schedule_id: scheduleId, group_id: groupId } = req.params;
-    const [user, group, schedule] = await Promise.all([
-      User.findOne({ where: { nickname: req.nickname } }),
-      Group.findByPk(groupId),
-      GroupSchedule.findByPk(scheduleId),
-    ]);
-
-    if (!schedule) {
-      return next(new ScheduleNotFoundError());
-    }
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (user.userId !== group.leader) {
-      return next(new EditPermissionError());
-    }
-
-    await schedule.destroy();
-
-    return res.status(204).json({ message: 'Successfully delete group schedule' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getSingleGroupSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupScheduleIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { schedule_id: scheduleId } = req.params;
-    const schedule = await GroupSchedule.findByPk(scheduleId);
-
-    if (!schedule) {
-      return next(new ScheduleNotFoundError());
-    }
-
-    return res.status(200).json(schedule);
   } catch (err) {
     return next(new ApiError());
   }
@@ -455,475 +298,14 @@ async function postGroupJoin(req, res, next) {
   }
 }
 
-async function getEventProposal(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupIdSchema(req.params);
-    const { error: queryError } = validateEventProposalSchema(req.query);
-
-    if (paramError || queryError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId } = req.params;
-    const group = await Group.findByPk(groupId);
-    const groupMembers = (await group.getUsers()).map((user) => user.userId);
-    const proposal = {};
-
-    /* eslint-disable no-restricted-syntax */
-    for (const date of Object.values(req.query)) {
-      let events = [];
-      const start = moment.utc(date).toDate();
-      const end = moment.utc(date).add(24, 'hours').add(-1, 's').toDate();
-      const {
-        nonRecurrenceSchedule: userNonRecEvent,
-        recurrenceSchedule: userRecEvent,
-        /* eslint-disable-next-line no-await-in-loop */
-      } = await PersonalSchedule.getSchedule(groupMembers, start, end);
-      events.push(parseEventDates(userNonRecEvent, userRecEvent));
-      const {
-        nonRecurrenceSchedule: groupNonRecEvent,
-        recurrenceSchedule: groupRecEvent,
-        /* eslint-disable-next-line no-await-in-loop */
-      } = await GroupSchedule.getSchedule([groupId], start, end);
-      events.push(parseEventDates(groupNonRecEvent, groupRecEvent));
-      events = events.flat(1);
-      events.sort((a, b) => a.startDateTime.getTime() - b.startDateTime.getTime());
-
-      // 결과값에서 9시~22시 사이의 값들을 먼저 추천할 수 있도록 정렬.
-      const result = eventProposal(events, start, end);
-      const filteredTimes = result.filter((event) => (
-        event.startDateTime.getTime() < (end.getTime() - 1000 * 60 * 60 * 2)
-          && event.endDateTime.getTime() > (start.getTime() + 1000 * 60 * 60 * 9)
-      ));
-      const remainingTimes = result.filter((event) => (
-        event.startDateTime.getTime() >= (end.getTime() - 1000 * 60 * 60 * 2)
-          || event.endDateTime.getTime() <= (start.getTime() + 1000 * 60 * 60 * 9)
-      ));
-      const sortedResult = filteredTimes.concat(remainingTimes);
-      proposal[date] = sortedResult;
-    }
-    return res.status(200).json(proposal);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function postGroupPost(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupIdSchema(req.params);
-    const { error: bodyError } = validatePostSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId } = req.params;
-    const [group, user] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    const { title, content } = req.body;
-    const post = await Post.create({ author: req.nickname, title });
-    await post.createPostDetail({ content });
-
-    await user.addPosts(post);
-    await group.addPosts(post);
-
-    return res.status(201).json({ message: 'Successfully created the post.' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getSinglePost(req, res, next) {
-  try {
-    const { error: paramError } = validatePostIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId } = req.params;
-    const [group, post] = await Promise.all([
-      Group.findByPk(groupId),
-      Post.findByPk(postId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    const { title, author } = post;
-    const { content } = (await post.getPostDetail()).dataValues;
-
-    return res.status(200).json({ author, title, content });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getGroupPosts(req, res, next) {
-  try {
-    const { error: paramError } = validateGroupIdSchema(req.params);
-    const { error: queryError } = validatePageSchema(req.query);
-
-    if (paramError || queryError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId } = req.params;
-    const group = await Group.findByPk(groupId);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    const { page } = req.query;
-    const pageSize = 7;
-
-    const startIndex = (page - 1) * pageSize;
-    const { rows } = await Post.findAndCountAll({
-      where: {
-        groupId,
-      },
-      include: [
-        {
-          model: PostDetail,
-          as: 'postDetail',
-        },
-      ],
-      offset: startIndex,
-      limit: pageSize,
-    });
-
-    const result = rows.map((post) => ({
-      title: post.title,
-      author: post.author,
-      createdAt: post.createdAt,
-      content: post.postDetail.content,
-    }));
-
-    return res.status(200).json(result);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function putGroupPost(req, res, next) {
-  try {
-    const { error: paramError } = validatePostIdSchema(req.params);
-    const { error: bodyError } = validatePostSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId } = req.params;
-    const [group, user, post] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-      Post.findByPk(postId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    if (user.userId !== post.userId) {
-      return next(new EditPermissionError());
-    }
-
-    const { title, content } = req.body;
-    await post.update({ title });
-    await PostDetail.update({ content }, { where: { postId } });
-
-    return res.status(200).json({ message: 'Successfully modified the post.' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function deleteGroupPost(req, res, next) {
-  try {
-    const { error: paramError } = validatePostIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId } = req.params;
-    const [group, user, post] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-      Post.findByPk(postId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    if (user.userId !== post.userId) {
-      return next(new EditPermissionError());
-    }
-
-    await post.destroy();
-
-    return res.status(204).end();
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getGroupList(req, res, next) {
-  try {
-    const { error: queryError } = validatePageSchema(req.query);
-    if (queryError) {
-      return next(new DataFormatError());
-    }
-
-    const { page } = req.query;
-    const pageSize = 9;
-    const startIndex = (page - 1) * pageSize;
-    const { rows } = await Group.findAndCountAll({
-      offset: startIndex,
-      limit: pageSize,
-    });
-    const result = rows.map((group) => ({
-      name: group.name,
-      member: group.member,
-    }));
-    return res.status(200).json(result);
-  } catch {
-    return next(new ApiError());
-  }
-}
-
-async function getPostComment(req, res, next) {
-  try {
-    const { error: paramError } = validatePostIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId } = req.params;
-    const [group, post] = await Promise.all([
-      Group.findByPk(groupId),
-      Post.findByPk(postId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    const comments = await post.getComments();
-
-    return res.status(200).json(comments);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getSingleComment(req, res, next) {
-  try {
-    const { error: paramError } = validateCommentIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId, comment_id: commentId } = req.params;
-    const [group, post, comment] = await Promise.all([
-      Group.findByPk(groupId),
-      Post.findByPk(postId),
-      Comment.findByPk(commentId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    if (!comment) {
-      return next(new CommentNotFoundError());
-    }
-
-    return res.status(200).json(comment);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function postComment(req, res, next) {
-  try {
-    const { error: paramError } = validatePostIdSchema(req.params);
-    const { error: bodyError } = validateCommentSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId } = req.params;
-    const [group, user, post] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-      Post.findByPk(postId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    const { content } = req.body;
-    const comment = await post.createComment({ content });
-    await user.addComments(comment);
-
-    return res.status(201).json({ message: 'Successfully created the comment.' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function putComment(req, res, next) {
-  try {
-    const { error: paramError } = validateCommentIdSchema(req.params);
-    const { error: bodyError } = validateCommentSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId, comment_id: commentId } = req.params;
-    const [group, user, post, comment] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-      Post.findByPk(postId),
-      Comment.findByPk(commentId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    if (!comment) {
-      return next(new CommentNotFoundError());
-    }
-
-    if (user.userId !== comment.userId) {
-      return next(new EditPermissionError());
-    }
-
-    const { content } = req.body;
-    await comment.update({ content });
-
-    return res.status(200).json({ message: 'Successfully modified the comment.' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function deleteComment(req, res, next) {
-  try {
-    const { error: paramError } = validateCommentIdSchema(req.params);
-    if (paramError) {
-      return next(new DataFormatError());
-    }
-
-    const { group_id: groupId, post_id: postId, comment_id: commentId } = req.params;
-    const [group, user, post, comment] = await Promise.all([
-      Group.findByPk(groupId),
-      User.findOne({ where: { nickname: req.nickname } }),
-      Post.findByPk(postId),
-      Comment.findByPk(commentId),
-    ]);
-
-    if (!group) {
-      return next(new GroupNotFoundError());
-    }
-
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    if (!post) {
-      return next(new PostNotFoundError());
-    }
-
-    if (!comment) {
-      return next(new CommentNotFoundError());
-    }
-
-    if (user.userId !== comment.userId) {
-      return next(new EditPermissionError());
-    }
-    await comment.destroy();
-
-    return res.status(204).end();
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
 module.exports = {
-  createGroup,
+  postGroup,
+  getGroupList,
   getGroupDetail,
-  deleteGroup,
   patchGroup,
+  deleteGroup,
   deleteGroupUser,
-  getGroupSchedule,
-  postGroupSchedule,
-  putGroupSchedule,
-  deleteGroupSchedule,
-  getSingleGroupSchedule,
   postInviteLink,
   getInvitation,
   postGroupJoin,
-  getEventProposal,
-  postGroupPost,
-  getGroupList,
-  getSinglePost,
-  getGroupPosts,
-  putGroupPost,
-  deleteGroupPost,
-  getPostComment,
-  getSingleComment,
-  postComment,
-  putComment,
-  deleteComment,
 };
