+++ src/controllers/user.js
@@ -1,25 +1,21 @@
-const moment = require('moment');
 const { Op } = require('sequelize');
 const bcrypt = require('bcrypt');
 
 // Model
 const User = require('../models/user');
+const UserGroup = require('../models/userGroup');
+const Group = require('../models/group');
 const PersonalSchedule = require('../models/personalSchedule');
-const GroupSchedule = require('../models/groupSchedule');
 
 // Error
 const {
   ApiError, DataFormatError,
-  UserNotFoundError, ScheduleNotFoundError, DuplicateUserError,
-  EditPermissionError,
+  UserNotFoundError, DuplicateUserError, UserIsLeaderError,
 } = require('../errors');
 
 // Validator
 const {
-  validateJoinSchema,
-  validateScheduleSchema,
-  validateScheduleIdSchema,
-  validateScheduleDateScehma,
+  validateJoinSchema, validateUserIdSchema,
 } = require('../utils/validators');
 
 async function getUserProfile(req, res, next) {
@@ -101,89 +97,67 @@ async function patchUserPassword(req, res, next) {
   }
 }
 
-async function getUserPersonalSchedule(req, res, next) {
+async function withdrawal(req, res, next) {
   try {
-    const { error: queryError } = validateScheduleDateScehma(req.query);
-    if (queryError) {
-      return next(new DataFormatError());
-    }
-
-    const user = await User.findOne({ where: { nickname: req.nickname } });
-    if (!user) {
-      return next(new UserNotFoundError());
-    }
-
-    const { startDateTime, endDateTime } = req.query;
-    const start = moment.utc(startDateTime).toDate();
-    const end = moment.utc(endDateTime).toDate();
-
-    const userEvent = await PersonalSchedule.getSchedule([user.userId], start, end);
-    const groups = (await user.getGroups()).map((group) => group.groupId);
-    if (groups.length) {
-      const groupEvent = await GroupSchedule.getSchedule(groups, start, end);
-      const event = {};
-      event.nonRecurrenceSchedule = [
-        ...userEvent.nonRecurrenceSchedule,
-        ...groupEvent.nonRecurrenceSchedule,
-      ];
-      event.recurrenceSchedule = [
-        ...userEvent.recurrenceSchedule,
-        ...groupEvent.recurrenceSchedule,
-      ];
-      return res.status(200).json(event);
-    }
-
-    return res.status(200).json(userEvent);
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function putUserSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateScheduleIdSchema(req.params);
-    const { error: bodyError } = validateScheduleSchema(req.body);
-
-    if (paramError || bodyError) {
-      return next(new DataFormatError());
-    }
-
-    const { schedule_id: scheduleId } = req.params;
-    const [user, schedule] = await Promise.all([
-      User.findOne({ where: { nickname: req.nickname } }),
-      PersonalSchedule.findByPk(scheduleId),
-    ]);
-
-    if (!schedule) {
-      return next(new ScheduleNotFoundError());
-    }
-
-    if (user.userId !== schedule.userId) {
-      return next(new EditPermissionError());
-    }
-
-    await PersonalSchedule.update(req.body, { where: { id: scheduleId } });
-    return res.status(201).json({ message: 'Successfully Modified.' });
-  } catch (err) {
-    return next(new ApiError());
-  }
-}
-
-async function getSingleUserSchedule(req, res, next) {
-  try {
-    const { error: paramError } = validateScheduleIdSchema(req.params);
+    const { error: paramError } = validateUserIdSchema(req.params);
     if (paramError) {
       return next(new DataFormatError());
     }
 
-    const { schedule_id: scheduleId } = req.params;
-    const schedule = await PersonalSchedule.findByPk(scheduleId);
+    const { user_id: userId } = req.params;
+    const [user, leader] = await Promise.all([
+      User.findByPk(userId),
+      Group.findOne({ where: { leader: userId } }),
+    ]);
 
-    if (!schedule) {
-      return next(new ScheduleNotFoundError());
+    if (!user) {
+      return next(new UserNotFoundError());
     }
 
-    return res.status(200).json(schedule);
+    if (leader) {
+      return next(new UserIsLeaderError());
+    }
+
+    await PersonalSchedule.destroy({ where: { userId } });
+    await user.destroy();
+
+    return res.status(204).json({ message: 'Successfully deleted' });
+  } catch (err) {
+    return next(new ApiError());
+  }
+}
+
+async function getUserSetup(req, res, next) {
+  try {
+    const user = await User.findOne({ where: { nickname: req.nickname } });
+    const groupList = await user.getGroups();
+    return res.status(200).json({ groupList });
+  } catch (err) {
+    return next(new ApiError());
+  }
+}
+
+async function updateUserSetUp(req, res, next) {
+  try {
+    const { error: paramError } = validateUserIdSchema(req.params);
+    if (paramError) {
+      return next(new DataFormatError());
+    }
+
+    const user = await User.findOne({ where: { nickname: req.nickname } });
+    const { updatedSharePersonalEvent } = req.body;
+
+    const updatePromises = updatedSharePersonalEvent.map(async (groupSetup) => {
+      const { groupId, sharePersonalEvent } = groupSetup;
+
+      await UserGroup.update(
+        { sharePersonalEvent },
+        { where: { groupId, userId: user.userId } },
+      );
+    });
+    await Promise.all(updatePromises);
+
+    return res.status(200).json({ message: 'Successfully update User Setup ' });
   } catch (err) {
     return next(new ApiError());
   }
@@ -194,7 +168,7 @@ module.exports = {
   getUserGroup,
   patchUserProfile,
   patchUserPassword,
-  getUserPersonalSchedule,
-  putUserSchedule,
-  getSingleUserSchedule,
+  withdrawal,
+  getUserSetup,
+  updateUserSetUp,
 };
