+++ src/routes/group.js
@@ -1,37 +1,45 @@
 const express = require('express');
 
+// Group
 const {
-  getGroupDetail, createGroup, patchGroup, deleteGroup,
+  postGroup, getGroupDetail, getGroupList, patchGroup, deleteGroup,
+  postInviteLink, getInvitation, postGroupJoin,
+} = require('../controllers/group');
+
+// Schedule
+const {
   getGroupSchedule, getSingleGroupSchedule,
   postGroupSchedule, putGroupSchedule, deleteGroupSchedule,
-  postInviteLink, getInvitation, postGroupJoin,
   getEventProposal,
-  getGroupList,
-  getGroupPosts, getSinglePost, postGroupPost, putGroupPost, deleteGroupPost,
-  getPostComment, getSingleComment, postComment, putComment, deleteComment,
-} = require('../controllers/group');
+} = require('../controllers/groupSchedule');
+
+// Feed
+const {
+  postGroupPost, getGroupPosts, getSinglePost, putGroupPost, deleteGroupPost,
+  postComment, getPostComment, getSingleComment, putComment, deleteComment,
+} = require('../controllers/feed');
 
 const router = express.Router();
 
 // Group
-router.post('/', createGroup);
+router.post('/', postGroup);
 router.get('/', getGroupList);
 router.get('/:group_id', getGroupDetail);
-router.delete('/:group_id', deleteGroup);
 router.patch('/:group_id', patchGroup);
+router.delete('/:group_id', deleteGroup);
 router.post('/:group_id/invite-link', postInviteLink);
 router.get('/invite-link/:inviteCode', getInvitation);
 router.post('/join/:inviteCode', postGroupJoin);
 
 // Schedule
-router.get('/:group_id/calendar', getGroupSchedule);
 router.post('/:group_id/calendar', postGroupSchedule);
+router.get('/:group_id/calendar', getGroupSchedule);
 router.get('/:group_id/calendar/:schedule_id', getSingleGroupSchedule);
 router.put('/:group_id/calendar/:schedule_id', putGroupSchedule);
 router.delete('/:group_id/calendar/:schedule_id', deleteGroupSchedule);
 router.get('/:group_id/proposal', getEventProposal);
 
-// Post
+// Feed
 router.post('/:group_id/post', postGroupPost);
 router.get('/:group_id/post', getGroupPosts);
 router.get('/:group_id/post/:post_id', getSinglePost);
