+++ src/routes/user.js
@@ -1,20 +1,18 @@
 const express = require('express');
 const { createToken } = require('../middleware/token');
 
+// User
 const {
   getUserGroup,
   patchUserProfile, patchUserPassword,
-  getUserPersonalSchedule, getSingleUserSchedule, putUserSchedule,
+  getUserSetup, updateUserSetUp,
 } = require('../controllers/user');
 
-const { 
-  getUserSetup, 
-  updateUserSetUp 
-} = require('../controllers/userSetup');
-
-const { 
-  postPersonalSchedule, deletePersonalSchedule 
-} = require('../controllers/calendar');
+// Schedule
+const {
+  getUserPersonalSchedule, getSingleUserSchedule,
+  postPersonalSchedule, putPersonalSchedule, deletePersonalSchedule,
+} = require('../controllers/personalSchedule');
 
 const { deleteGroupUser } = require('../controllers/group');
 
@@ -32,6 +30,6 @@ router.get('/userSetup', getUserSetup);
 router.get('/calendar', getUserPersonalSchedule);
 router.get('/calendar/:schedule_id', getSingleUserSchedule);
 router.post('/calendar', postPersonalSchedule);
-router.put('/calendar/:schedule_id', putUserSchedule);
+router.put('/calendar/:schedule_id', putPersonalSchedule);
 router.delete('/calendar/:schedule_id', deletePersonalSchedule);
 module.exports = router;
