+++ src/utils/validators.js
@@ -39,7 +39,7 @@ const groupIdSchema = Joi.object({
   group_id: Joi.number().min(0).required(),
 });
 
-const scheduleDateScehma = Joi.object({
+const scheduleDateSchema = Joi.object({
   startDateTime: Joi.date(),
   endDateTime: Joi.date(),
 });
@@ -132,7 +132,7 @@ module.exports = {
   validateScheduleIdSchema: validator(scheduleIdSchema),
   validateGroupScheduleIdSchema: validator(groupScheduleIdSchema),
   validateGroupScheduleSchema: validator(groupScheduleSchema),
-  validateScheduleDateScehma: validator(scheduleDateScehma),
+  validateScheduleDateSchema: validator(scheduleDateSchema),
   validateUserScheduleSchema: validator(userScheduleSchema),
   validateYYYYMMDateSchema: validator(yearMonthScehma),
   validateYYYYMMDDDateSchema: validator(yearMonthDayScehma),
